Develop m2fixes

See merge request xsokolar/pv-168-project-share-car-rider!21

+++ src/main/java/cz/muni/fi/pv168/project/database/entity/RideEntity.java
@@ -2,23 +2,24 @@ package cz.muni.fi.pv168.project.database.entity;
 
 import cz.muni.fi.pv168.project.model.Currency;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
 public record RideEntity(
         Integer id,
-        float distance,
+        BigDecimal distance,
         LocalDateTime dateTime,
-        float price,
+        BigDecimal price,
         Currency originalCurrency,
         int categoryId,
         int passengersCount) {
 
     public RideEntity(
             Integer id,
-            float distance,
+            BigDecimal distance,
             LocalDateTime dateTime,
-            float price,
+            BigDecimal price,
             Currency originalCurrency,
             int categoryId,
             int passengersCount) {
@@ -32,9 +33,9 @@ public record RideEntity(
     }
 
     public RideEntity(
-            float distance,
+            BigDecimal distance,
             LocalDateTime dateTime,
-            float price,
+            BigDecimal price,
             Currency originalCurrency,
             int categoryId,
             int passengersCount) {
